requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Security
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Validation
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9
